set DataSource for Writing Domain Configuration
