+++ core/src/main/java/com/snim/configuration/DataSourcesConfiguration.java
@@ -1,0 +1,11 @@
+package com.snim.configuration;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+@Configuration
+@Import({
+        com.snim.core.writing.DataSourceConfiguration.class
+})
+public class DataSourcesConfiguration {
+}
