--- core/src/main/java/com/snim/configuration/jdbc/MysqlProperties.java
@@ -1,6 +1,0 @@
-package com.snim.configuration.jdbc;
-
-import org.springframework.boot.autoconfigure.jdbc.DataSourceProperties;
-
-public class MysqlProperties extends DataSourceProperties {
-}
